+++ changelog.md
@@ -2,7 +2,23 @@
 
 All notable changes to `Elasticsearch` will be documented in this file.
 
-## Version 0.x.x
+## Version 0.6.0
+### Breaking changes
+
+Any value type aggregation (sum, value_count, avg, ...) which returned a `SimpleValue` object having `value()` method
+and `value` property on it's fluent class, no longer turn `SimpleValue`
+
+```php
+// Old 
+$result->aggregation->total->value();
+
+// New 
+$result->aggregation->total;
+``` 
+
+## Version 0.5.2
+## Version 0.5.1
+## Version 0.5.0
 ### Added
 
 ### Fixes